Change the if checks to help add more info
Changed the if checks so that it would include things like "Inner Damage" or "Damage Per Hit" vs "Damage per Hit".

There is also a weird thing with Pele that her Outer Damage includes "isrankx" in it, so I changed the num(s) function to ensure that it will not error out.